Extract out clean writes and visibility utilities
