+++ src/main/kotlin/com/slack/auto/value/kotlin/KotlinClass.kt
@@ -64,8 +64,7 @@ public data class KotlinClass(
   val staticConstants: List<PropertySpec>,
 ) {
   @Suppress("LongMethod", "ComplexMethod")
-  @OptIn(ExperimentalPathApi::class)
-  public fun writeTo(dir: String, messager: Messager) {
+  public fun toTypeSpec(messager: Messager): TypeSpec {
     val typeBuilder = TypeSpec.classBuilder(name)
       .addModifiers(DATA)
       .addAnnotations(classAnnotations)
@@ -217,69 +216,7 @@ public data class KotlinClass(
       typeBuilder.addType(companionObjectBuilder.build())
     }
 
-    val file = File(dir).toPath()
-    val outputPath = FileSpec.get(packageName, typeBuilder.build())
-      .writeToLocal(file)
-    val text = outputPath.readText()
-    // Post-process to remove any kotlin intrinsic types
-    // Is this wildly inefficient? yes. Does it really matter in our cases? nah
-    var prevWasBlank = false
-    outputPath.writeText(
-      text
-        .lineSequence()
-        .filterNot { it in INTRINSIC_IMPORTS }
-        .mapNotNull {
-          if (it.trimStart().startsWith("public ")) {
-            prevWasBlank = false
-            val indent = it.substringBefore("public ")
-            it.removePrefix(indent).removePrefix("public ").prependIndent(indent)
-          } else if (it.isKotlinPackageImport) {
-            // Ignore kotlin implicit imports
-            null
-          } else if (it.isBlank()) {
-            if (prevWasBlank) {
-              null
-            } else {
-              prevWasBlank = true
-              it
-            }
-          } else {
-            prevWasBlank = false
-            it
-          }
-        }
-        .joinToString("\n")
-    )
-  }
-
-  /** Best-effort checks if the string is an import from `kotlin.*` */
-  @Suppress("MagicNumber")
-  private val String.isKotlinPackageImport: Boolean get() = startsWith("import kotlin.") &&
-    // Looks like a class
-    // 14 is the length of `import kotlin.`
-    get(14).isUpperCase() &&
-    // Exclude if it's importing a nested element
-    '.' !in removePrefix("import kotlin.")
-
-  private fun FileSpec.writeToLocal(directory: Path): Path {
-    require(Files.notExists(directory) || Files.isDirectory(directory)) {
-      "path $directory exists but is not a directory."
-    }
-    var srcDirectory = directory
-    if (packageName.isNotEmpty()) {
-      for (packageComponent in packageName.split('.').dropLastWhile { it.isEmpty() }) {
-        srcDirectory = srcDirectory.resolve(packageComponent)
-      }
-    }
-
-    Files.createDirectories(srcDirectory)
-
-    val outputPath = srcDirectory.resolve("$name.kt")
-    OutputStreamWriter(
-      Files.newOutputStream(outputPath),
-      StandardCharsets.UTF_8
-    ).use { writer -> writeTo(writer) }
-    return outputPath
+    return typeBuilder.build()
   }
 
   // Public for extension
